Move `create_app_runner`, `wait_for_restart`, and `wait_for_shutdown` to separate `process_utils` module.
Required to avoid circular imports.
