+++ virtool/app.py
@@ -1,13 +1,10 @@
 import asyncio
 import logging
-import os
-import sys
 
 import aiohttp.web
 import aiojobs
 import aiojobs.aiohttp
 
-import virtool.config
 import virtool.db.core
 import virtool.db.migrate
 import virtool.db.utils
@@ -31,48 +28,9 @@ import virtool.software.db
 import virtool.startup
 import virtool.utils
 import virtool.version
+from virtool.process_utils import create_app_runner, wait_for_restart, wait_for_shutdown, logger
 
 logger = logging.getLogger(__name__)
-
-
-async def on_shutdown(app: aiohttp.web.Application):
-    """
-    A function called when the app is shutting down.
-
-    :param app: the app object
-    :type app: :class:`aiohttp.web.Application`
-
-    """
-    logger.debug("Shutting down")
-
-    try:
-        await app["client"].close()
-    except KeyError:
-        pass
-
-    try:
-        await app["dispatcher"].close()
-    except KeyError:
-        pass
-
-    try:
-        app["executor"].shutdown(wait=True)
-    except KeyError:
-        pass
-
-    try:
-        app["process_executor"].shutdown(wait=True)
-    except KeyError:
-        pass
-
-    scheduler = virtool.startup.get_scheduler_from_app(app)
-    await scheduler.close()
-
-    try:
-        app["redis"].close()
-        await app["redis"].wait_closed()
-    except KeyError:
-        pass
 
 
 def create_app(config):
@@ -122,40 +80,6 @@ def create_app(config):
     return app
 
 
-async def create_app_runner(app: aiohttp.web.Application, host: str, port: int) -> aiohttp.web.AppRunner:
-    """
-    Create an :class:`aiohttp.web.AppRunner` to allow customization of signal handlers.
-
-    The basic :func:`aiohttp.web.run_app` sets up handlers for `SIGINT` and `SIGTERM` that can interfere with Virtool
-    code such as that for restarting the server after software update. This custom runner allows handling of signals
-    as well as restart and shutdown events from users.
-
-    https://docs.aiohttp.org/en/stable/web_advanced.html#application-runners
-
-    :param app: the application
-    :param host: the host to listen on
-    :param port: the port to listen on
-    :return: a custom :class:`~aiohttp.web.AppRunner`
-
-    """
-    runner = aiohttp.web.AppRunner(app)
-
-    await runner.setup()
-
-    site = aiohttp.web.TCPSite(runner, host, port)
-
-    try:
-        await site.start()
-    except OSError as err:
-        if err.args[0] == 48:
-            logger.fatal(f"Could not bind address {(host, port)}")
-            sys.exit(1)
-
-    logger.info(f"Listening at http://{host}:{port}")
-
-    return runner
-
-
 async def run_app(config):
     app = create_app(config)
 
@@ -174,41 +98,41 @@ async def run_app(config):
         task.cancel()
 
 
-async def wait_for_restart(runner: aiohttp.web.AppRunner, events: dict):
+async def on_shutdown(app: aiohttp.web.Application):
     """
-    Wait for the shutdown event and restart if it is encountered.
+    A function called when the app is shutting down.
 
-    Restart is accomplished using :func:`os.execl` or :func:`os.execv` after cleaning up the `runner`.
-
-    :param runner: the :class:`~aiohttp.web.AppRunner` returned by :func:`.create_app_runner`
-    :param events: a dict containing the `restart` and `shutdown` :class:`~asyncio.Event` objects
+    :param app: the app object
+    :type app: :class:`aiohttp.web.Application`
 
     """
-    await events["restart"].wait()
-    await asyncio.sleep(0.5)
-    await runner.cleanup()
+    logger.debug("Shutting down")
 
-    exe = sys.executable
+    try:
+        await app["client"].close()
+    except KeyError:
+        pass
 
-    if exe.endswith("python") or "python3" in exe:
-        os.execl(exe, exe, *sys.argv)
-        return
+    try:
+        await app["dispatcher"].close()
+    except KeyError:
+        pass
 
-    if exe.endswith("run"):
-        os.execv(exe, sys.argv)
-        return
+    try:
+        app["executor"].shutdown(wait=True)
+    except KeyError:
+        pass
 
-    raise SystemError("Could not determine executable type")
+    try:
+        app["process_executor"].shutdown(wait=True)
+    except KeyError:
+        pass
 
+    scheduler = virtool.startup.get_scheduler_from_app(app)
+    await scheduler.close()
 
-async def wait_for_shutdown(runner: aiohttp.web.AppRunner, events: dict):
-    """
-    Wait for the shutdown event and terminate if it is encountered.
-
-    :param runner: the :class:`~aiohttp.web.AppRunner` returned by :func:`.create_app_runner`
-    :param events: a dict containing the `restart` and `shutdown` :class:`~asyncio.Event` objects
-
-    """
-    await events["shutdown"].wait()
-    await asyncio.sleep(0.5)
-    await runner.cleanup()
+    try:
+        app["redis"].close()
+        await app["redis"].wait_closed()
+    except KeyError:
+        pass
